refactor: Move exam finish event to HTTP
- remove integration of `Score` model and `finishExam` handler function in `socket.exam.js`.
- refactor `finishExam` to utilize HTTP request/response objects instead of socket event emission format for exam processing in `examController.js`.
- integrate `finishExam` in its respective API route endpoint in `examRoutes.js`.
