+++ backend/controller/socket/helpers/socket.exam.js
@@ -1,6 +1,5 @@
 import Exam from "../../../models/quizzes/examModel.js";
 import Quiz from "../../../models/quizzes/quizModel.js";
-import Score from "../../../models/quizzes/scoreModel.js";
 import Classroom from "../../../models/classrooms/classroomModel.js";
 
 export const createExam = async (context, data) => {
@@ -105,105 +104,6 @@ export const updateExam = async (context, data) => {
 
     context.emitEvent("sender", "error", {
       message: `Error updating exam ${data?.receiverId}:  ${error.message}`,
-    });
-  }
-};
-
-export const finishExam = async (context, data) => {
-  try {
-    const { senderId, receiverId } = data;
-
-    const examFound = await Exam.findOne({ studentId: senderId });
-
-    if (!examFound) {
-      throw new Error("Exam not found");
-    }
-
-    const examQuestions = examFound?.shuffledQuestions;
-
-    const examIsValid = examFound?.examFinish?.getTime() - Date.now();
-
-    if (!examIsValid) {
-      throw new Error("Exam has expired");
-    }
-
-    await Exam.findByIdAndUpdate(
-      examFound?._id,
-      { isInProgress: false },
-      { new: true }
-    );
-
-    const quizFound = await Quiz.findOne({ quizId: receiverId });
-
-    if (!quizFound) {
-      throw new Error("Cannot evaluate exam, no matching quiz found");
-    }
-
-    let payloadScore = 0;
-
-    let userChoices = [];
-
-    for (let i = 0; i < quizFound.questions.length; i++) {
-      const q = quizFound.questions[i];
-
-      if (q.answer === examFound.answers[i]) {
-        payloadScore += 1;
-      }
-
-      userChoices[i] = {
-        userAnswer: examFound?.answers?.[i],
-        correctAnswer: q.answer,
-      };
-    }
-
-    const scoreFound = await Score.findOne({
-      user: senderId,
-      quiz: quizFound?._id,
-    });
-
-    let scorePayload;
-
-    if (scoreFound) {
-      const updatedScore = await Score.findByIdAndUpdate(
-        scoreFound?._id,
-        {
-          $set: {
-            "examFeedback.userChoices": userChoices,
-            "examFeedback.randomizedQuestions": examQuestions,
-            highScore: Math.max(scoreFound.highScore, payloadScore),
-          },
-        },
-        { new: true }
-      );
-
-      scorePayload = updatedScore;
-    } else {
-      const newScore = new Score({
-        user: senderId,
-        quiz: quizFound?._id,
-        examFeedback: {
-          userChoices,
-          randomizedQuestions: examQuestions,
-        },
-        highScore: payloadScore,
-      });
-
-      const updatedScore = await newScore.save();
-      scorePayload = updatedScore;
-    }
-
-    const examPayload = { examId: examFound?._id, scorePayload };
-
-    await Exam.findByIdAndDelete({ _id: examFound?._id });
-
-    context.emitEvent("sender", "exam finished", examPayload);
-  } catch (error) {
-    console.error(
-      `Error finishing exam ${data?.receiverId}:  ${error.message}`
-    );
-
-    context.emitEvent("sender", "error", {
-      message: `Error finishing exam ${data?.receiverId}:  ${error.message}`,
     });
   }
 };
