+++ backend/routes/quizzes/examRoutes.js
@@ -4,9 +4,11 @@ const examRoutes = Router();
 import {
   getExam,
   getExamFeedback,
+  finishExam,
 } from "../../controller/quizzes/examController.js";
 import { protect } from "../../middleware/authMiddleware.js";
 
+examRoutes.post("/", protect, finishExam);
 examRoutes.get("/", protect, getExam);
 examRoutes.get("/feedback/:quizId", protect, getExamFeedback);
